backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.1
pandas>=2.2.2
openpyxl>=3.1.2
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.1.1
